refactor: errorCode location
