+++ src/constants/error/auth/auth-error-message.constant.ts
@@ -1,4 +1,4 @@
-import { AUTH_ERROR_CODE } from '@src/apis/auth/constants/auth-error-code.constant';
+import { AUTH_ERROR_CODE } from '@src/constants/error/auth/auth-error-code.constant';
 import { ErrorMessage } from '@src/types/type';
 
 export const AUTH_ERROR_MESSAGE: ErrorMessage<typeof AUTH_ERROR_CODE> = {
