+++ src/constants/error/error-code.constant.ts
@@ -1,0 +1,9 @@
+import { AUTH_ERROR_CODE } from '@src/constants/error/auth/auth-error-code.constant';
+import { COMMON_ERROR_CODE } from '@src/constants/error/common/common-error-code.constant';
+import { USER_ERROR_CODE } from '@src/constants/error/users/user-error-code.constant';
+
+export const ERROR_CODE = {
+  ...COMMON_ERROR_CODE,
+  ...USER_ERROR_CODE,
+  ...AUTH_ERROR_CODE,
+} as const;
