+++ src/types/type.ts
@@ -2,6 +2,8 @@ export type Nullable<T> = {
   [P in keyof T]: T[P] | null;
 };
 
+export type ErrorCode = Record<string, number>;
+
 export type ErrorMessage<T extends Record<string, number>> = Required<{
   [key in T[keyof T]]: string;
 }>;
